requirements.txt
streamlit==1.42.0
pandas==2.2.3
plotly==6.0.0
google-generativeai==0.8.4
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
Flask-CORS==5.0.0
